simplified sanitation
