+++ copy_tool/Cargo.toml
@@ -4,8 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-clippers = "0.1.2"
+#clippers = "0.1.2"
 crossterm = "0.27.0"
-ratatui = "0.26.1"
 regex = "1.10.4"
 serial2 = "0.2.20"
+notify-rust = "4.11.0"
